Use web worker
