+++ typescript/utils/image_utils.ts
@@ -15,10 +15,7 @@
  * limitations under the License.
  */
 
-import {QuantizerCelebi} from '../quantize/quantizer_celebi.js';
-import {Score} from '../score/score.js';
-
-import {argbFromRgb} from './color_utils.js';
+import {rankedColorsFromImageBytes} from './image_utils_converter.js';
 
 /**
  * Get the source color from an image.
@@ -29,7 +26,8 @@ import {argbFromRgb} from './color_utils.js';
 export async function sourceColorFromImage(image: HTMLImageElement) {
   // Convert Image data to Pixel Array
   const imageBytes = await new Promise<Uint8ClampedArray>((resolve, reject) => {
-    const canvas = document.createElement('canvas');
+    const element = document.createElement('canvas');
+    const canvas = 'OffscreenCanvas' in window ? element.transferControlToOffscreen() : element;
     const context = canvas.getContext('2d');
     if (!context) {
       reject(new Error('Could not get canvas context'));
@@ -38,6 +36,7 @@ export async function sourceColorFromImage(image: HTMLImageElement) {
     const callback = () => {
       canvas.width = image.width;
       canvas.height = image.height;
+      // @ts-ignore
       context.drawImage(image, 0, 0);
       let rect = [0, 0, image.width, image.height];
       const area = image.dataset['area'];
@@ -48,6 +47,7 @@ export async function sourceColorFromImage(image: HTMLImageElement) {
         });
       }
       const [sx, sy, sw, sh] = rect;
+      // @ts-ignore
       resolve(context.getImageData(sx, sy, sw, sh).data);
     };
     if (image.complete) {
@@ -57,23 +57,23 @@ export async function sourceColorFromImage(image: HTMLImageElement) {
     }
   });
 
-  // Convert Image data to Pixel Array
-  const pixels: number[] = [];
-  for (let i = 0; i < imageBytes.length; i += 4) {
-    const r = imageBytes[i];
-    const g = imageBytes[i + 1];
-    const b = imageBytes[i + 2];
-    const a = imageBytes[i + 3];
-    if (a < 255) {
-      continue;
-    }
-    const argb = argbFromRgb(r, g, b);
-    pixels.push(argb);
+  let ranked: number[];
+
+  if (window.Worker) {
+    const worker = new Worker(new URL('./image_utils_worker.js', import.meta.url), {type: 'module'});
+
+    worker.postMessage(imageBytes);
+  
+    ranked = await new Promise((resolve) => {
+      worker.onmessage = (event) => {
+        const ranked = event.data;
+        resolve(ranked);
+      };
+    });
+  } else {
+    ranked = rankedColorsFromImageBytes(imageBytes);
   }
 
-  // Convert Pixels to Material Colors
-  const result = QuantizerCelebi.quantize(pixels, 128);
-  const ranked = Score.score(result);
   const top = ranked[0];
   return top;
 }
