+++ typescript/utils/image_utils_worker.ts
@@ -1,0 +1,24 @@
+/**
+ * @license
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {rankedColorsFromImageBytes} from './image_utils_converter.js';
+
+self.onmessage = (event) => {
+  const imageBytes = event.data;
+  const ranked = rankedColorsFromImageBytes(imageBytes);
+  self.postMessage(ranked);
+}
